Updates from v1.11.6
Some changes are not in the same files as code was moved during large refactoring efforts:
- src/pke/include/cryptoobject.h/cryptoobject.h:168
- src/pke/include/schemebase/rlwe-cryptoparameters.h:385
- src/pke/include/cryptoobject.h:169 GetEncryptionAlgorithm() is no longer a valid method, switched to GetScheme()

New changes to met the spirit of v1.11.6 changes
- src/pke/include/ciphertext.h:554 m_hopslevel type changed from size_t to uint32_t for WASM issues.

CRYP-583

+++ src/core/include/math/hal/bigintdyn/ubintdyn.h
@@ -931,7 +931,7 @@ class ubint : public lbcrypto::BigIntegerInterface<ubint<limb_t>> {
       result = (T)m_value[0];
     } else {
       // Case where limb_t is less bits than uint64_t
-      size_t msbTest = sizeof(T) * 8;
+      uint32_t msbTest = sizeof(T) * 8;
       if (msbTest > m_MSB) {
         msbTest = m_MSB;
       }
@@ -1229,7 +1229,7 @@ class ubint : public lbcrypto::BigIntegerInterface<ubint<limb_t>> {
 
  private:
   // variable that stores the MOST SIGNIFICANT BIT position in the
-  size_t m_MSB;
+  uint32_t m_MSB;
 
   // variable to store the bitlength of the limb data type.
   static const usint m_limbBitLength;
